fix(tests): unbreak component-tests build — stop a client modal pulling the appRouter (sharp) into the browser bundle
The `component` Vitest suite (`pnpm test:component`) has been failing to BUILD
(not assert) on `main`, so it is silently dark for everyone:

    ✘ No loader is configured for ".node" files:
      node_modules/.pnpm/sharp@0.32.6/.../sharp-linux-x64.node

Root cause (import chain, confirmed via an esbuild metafile trace over all
`*.browser.test.tsx` entries — Vite's dep scanner crawls the whole set):

  *.browser.test.tsx
    → …client modals… → src/components/CosmeticShop/CosmeticShopItemPreviewModal.tsx
    → ~/pages/moderator/cosmetic-store/cosmetics (a Next PAGE)
    → getServerSideProps → ~/server/utils/server-side-helpers
    → ~/server/routers/index.ts (the tRPC appRouter)
    → ~/server/routers/app-listings.router.ts
    → ~/server/services/blocks/app-listing-assets.service.ts  → await import('sharp')

`CosmeticShopItemPreviewModal` (a client component) imported `CosmeticPreview`
from the moderator cosmetics *page* module. A page's `getServerSideProps`
transitively pulls the entire server appRouter into whatever bundles it — a
long-standing client/server layering leak. It was harmless (bundled the whole
server as plain JS) until #2865 (W13 AppListing asset pipeline) registered
`app-listings.router` into the appRouter, and that router `await import('sharp')`s
a native `.node` addon → esbuild has no loader → the whole component build dies.
(#2838, the suspected card PR, is NOT in the chain — its files never reach sharp.)

Fix (root cause, not a bundler stub): extract the pure-client `CosmeticPreview`
component out of the page into its own client module
`~/components/CosmeticShop/CosmeticPreview.tsx`, and point both the page and the
modal at it. The modal no longer imports a page module, so the appRouter (and
sharp) drop out of the client/test bundle entirely. Verified via the metafile
trace (sharp/app-listing-assets no longer reachable) and by running the suite:
component-tests now BUILDS and RUNS — 35 files / 344 tests pass (was: 0, hard
build failure). Unused-after-extraction imports were pruned from the page; no
tests were skipped or disabled. Unit project is a separate pre-existing failure,
untouched (no unit test references the changed files).

Follow-up (not done here): a lint guard forbidding client components from
importing `~/pages/**` would catch this class of leak deterministically.
