+++ src/components/CosmeticShop/CosmeticShopItemPreviewModal.tsx
@@ -23,7 +23,7 @@ import {
 import { useDialogContext } from '~/components/Dialog/DialogProvider';
 import { dialogStore } from '~/components/Dialog/dialogStore';
 import { useCurrentUser } from '~/hooks/useCurrentUser';
-import { CosmeticPreview } from '~/pages/moderator/cosmetic-store/cosmetics';
+import { CosmeticPreview } from '~/components/CosmeticShop/CosmeticPreview';
 import type { CosmeticShopItemGetById } from '~/types/router';
 import { showSuccessNotification } from '~/utils/notifications';
 import { getDisplayName } from '~/utils/string-helpers';
